Move events -> pi-events
Prevent name collision
