+++ src/agent2/pi-events.ts
